kr_3/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
PyJWT
slowapi
